Audit the comments across the repo
The capture engine carried the most and gave up the most: 40 lines, nearly all of
them where a named test already holds the fact. A comment cannot out-protect an
assertion, and these tests say what they guard —
testFullWidthCardHasSymmetricPaddingNotAGiantTrailingGap,
testRowFillsWidthFromNestedSpacer, testMixedHeightRowOrdersLeftToRightByCentre,
testLiteralWhiteTextDoesNotBindABackgroundToken, testSalePillFillSurvivesCapture,
testTwoDimensionalRowReflectsThumbnailInfoColumnStepper — so the prose beside them
was a second, weaker copy.

Two facts stated twice now live once, at the declaration: the system font's family
name is not Figma-loadable, and a text colour binds only to a text-role token.

Two became names. The containment tree sorted by area with a comment explaining
that a parent must be placed before its children; the sort is
parentsBeforeChildren now. TweakableUITests described three suites' worth of
coverage that moved to DemoTests and held a single device-crash test, so it is
SimulatedDeviceUITests and the stale header is gone with its dead catalog helper.

Seven MARK banners went, per the house rule against decorative dividers, along
with the version snapshots — a v5 scramble and a v7 corruption nobody can look up
— and a "see the plan" pointer.

What stayed is what nothing else can hold: SwiftUI's private DisplayList
internals, the AttributeGraph call signature, Figma's own behaviour where a bound
style detaches when a style-owned property is written and addMode throws on a free
plan, and the measured reason a soft fill cannot carry a selected state.

Tests: unit 123/123 + hosted 11/11 green, UI target builds (local xcodebuild)
